na frisst mono das?

+++ microcontroller/src-host/src/liblap/proto_mood.h
@@ -4,6 +4,6 @@
 // functions
 unsigned char mood_set_mood( can_addr dst,unsigned char mood, unsigned char val1, unsigned char val2, unsigned char val3, unsigned char val4 );
 
-void mood_get_mood(can_addr dst,uint8_t mood,uint8_t * val1,uint8_t * val2,uint8_t * val3,uint8_t * val4);
+void mood_get_mood(can_addr dst,uint8_t mood,int * val1,int * val2,int * val3,int * val4);
 
 #endif
